requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
mysql-connector==2.2.9
mysqlclient==2.2.4
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
xlrd==2.0.1
